+++ Changelog.md
@@ -14,6 +14,7 @@ Bugfixes:
  * Standalone Assembly: Do not ignore input after closing brace of top level block.
  * Standard JSON: catch errors properly when invalid "sources" are passed
  * Type Checker: Properly warn when using ``_offset`` and ``_slot`` for constants in inline assembly.
+ * Commandline interface: throw error if option is unknown
 
 ### 0.4.20 (2018-02-14)